requirements.txt
APScheduler==3.6.3
asgiref==3.5.2
backports.zoneinfo==0.2.1
black==22.3.0
cachetools==4.2.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
python-dotenv==0.20.0
python-telegram-bot==13.12
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
